# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PVT OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/global/procedure_qr/{module_id}/{uuid}": {
            "get": {
                "tags": [
                    "TRÁMITES"
                ],
                "summary": "TRÁMITE DE ACUERDO AL MÓDULO Y UUID SOLICITADO",
                "operationId": "getProcedureQRByModuleAndUuid",
                "parameters": [
                    {
                        "name": "module_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 6
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "example": "cc2f6a58-9ea8-46f3-94df-c3e61aa3bbcc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}